+++ src/interfaces/response.ts
@@ -2,8 +2,8 @@ export interface AnimeSubsApiResponse {
   status: number;
   message: string;
   episode_url?: {
-    player: string;
-    url: string;
+    player: string | null;
+    url: string | null;
   }[];
   episode_next_url?: string | null;
 }
